+++ examples/hours.rs
@@ -7,16 +7,16 @@ fn main() {
     //Source https://en.wikipedia.org/wiki/Wikipedia:Size_of_Wikipedia
     let data = [
         (
-            UnixTime::from_ymd_hms(chrono::Utc, day1, 12, 59, 59),
+            UnixTime::from_ymd_hms(&chrono::Utc, day1, 12, 59, 59),
             3144000,
         ),
-        (UnixTime::from_ymd_hms(chrono::Utc, day2, 1, 0, 0), 3518000),
-        (UnixTime::from_ymd_hms(chrono::Utc, day2, 2, 2, 1), 3835000),
+        (UnixTime::from_ymd_hms(&chrono::Utc, day2, 1, 0, 0), 3518000),
+        (UnixTime::from_ymd_hms(&chrono::Utc, day2, 2, 2, 1), 3835000),
         (
-            UnixTime::from_ymd_hms(chrono::Utc, day2, 3, 59, 59),
+            UnixTime::from_ymd_hms(&chrono::Utc, day2, 3, 59, 59),
             2133000,
         ),
-        (UnixTime::from_ymd_hms(chrono::Utc, day2, 6, 1, 1), 4133000),
+        (UnixTime::from_ymd_hms(&chrono::Utc, day2, 6, 1, 1), 4133000),
     ];
 
     let mut s = poloto::plot(
