+++ examples/years_fmt.rs
@@ -1,22 +1,24 @@
-use poloto::num::timestamp::UnixTime;
+use poloto::num::timestamp::{UnixTime, UnixTimeContext};
 use poloto::prelude::*;
 // PIPE me to a file!
 fn main() {
+    let timezone = &chrono::Utc;
+
     //Source https://en.wikipedia.org/wiki/Wikipedia:Size_of_Wikipedia
     let data = [
-        (UnixTime::from_year(chrono::Utc, 2010), 3144000),
-        (UnixTime::from_year(chrono::Utc, 2011), 3518000),
-        (UnixTime::from_year(chrono::Utc, 2012), 3835000),
-        (UnixTime::from_year(chrono::Utc, 2013), 4133000),
-        (UnixTime::from_year(chrono::Utc, 2014), 4413000),
-        (UnixTime::from_year(chrono::Utc, 2015), 4682000),
-        (UnixTime::from_year(chrono::Utc, 2016), 5045000),
-        (UnixTime::from_year(chrono::Utc, 2017), 5321200),
-        (UnixTime::from_year(chrono::Utc, 2018), 5541900),
-        (UnixTime::from_year(chrono::Utc, 2019), 5773600),
-        (UnixTime::from_year(chrono::Utc, 2020), 5989400),
-        (UnixTime::from_year(chrono::Utc, 2021), 6219700),
-        (UnixTime::from_year(chrono::Utc, 2022), 0), //To complete our histogram, we manually specify when 2021 ends.
+        (UnixTime::from_year(timezone, 2010), 3144000),
+        (UnixTime::from_year(timezone, 2011), 3518000),
+        (UnixTime::from_year(timezone, 2012), 3835000),
+        (UnixTime::from_year(timezone, 2013), 4133000),
+        (UnixTime::from_year(timezone, 2014), 4413000),
+        (UnixTime::from_year(timezone, 2015), 4682000),
+        (UnixTime::from_year(timezone, 2016), 5045000),
+        (UnixTime::from_year(timezone, 2017), 5321200),
+        (UnixTime::from_year(timezone, 2018), 5541900),
+        (UnixTime::from_year(timezone, 2019), 5773600),
+        (UnixTime::from_year(timezone, 2020), 5989400),
+        (UnixTime::from_year(timezone, 2021), 6219700),
+        (UnixTime::from_year(timezone, 2022), 0), //To complete our histogram, we manually specify when 2021 ends.
     ];
 
     let xname = poloto::fmt::name_ext(|w, ([min, max], xs), _| {
@@ -30,8 +32,8 @@ fn main() {
         "title",
         xname,
         "yname",
-        UnixTime::default_ctx()
-            .with_tick_fmt(|w, v, _b, s| write!(w, "{} yr", v.dynamic_format(&chrono::Utc, s))),
+        UnixTimeContext::new(timezone)
+            .with_tick_fmt(|w, v, _b, s| write!(w, "{} yr", v.dynamic_format(timezone, s))),
         i128::default_ctx()
             .with_no_dash()
             .with_marker(0)
