+++ src/plotnum/ext.rs
@@ -1,4 +1,3 @@
-
 //!
 //! Contains the [`PlotNumContextExt`] trait that provides adaptor functions modifying a [`PlotNumContext`].
 //!
